drm/mediatek: Fix integer truncation warning
gcc throws a warning due to a 64-bit integer being implicitly truncated
to 32 bits.

In file included from drivers/gpu/drm/mediatek/mtk_dsi.c:15:
drivers/gpu/drm/mediatek/mtk_dsi.c: In function ‘mtk_dsi_reset_all’:
drivers/gpu/drm/mediatek/mtk_dsi.c:290:27: error: conversion from ‘long unsigned int’ to ‘unsigned int’ changes value from ‘18446744073675997183’ to ‘4261412863’ [-Werror=overflow]
  290 |       MMSYS_SW_RST_DSI_B, ~MMSYS_SW_RST_DSI_B);
include/linux/regmap.h:75:42: note: in definition of macro ‘regmap_update_bits’
   75 |  regmap_update_bits_base(map, reg, mask, val, NULL, false, false)
      |                                          ^~~

Fix it with a cast.

Fixes: d5b33be3b208 ("drm/mediatek: add dsi module reset driver")

+++ drivers/gpu/drm/mediatek/mtk_dsi.c
@@ -287,7 +287,7 @@ static void mtk_dsi_reset_all(struct mtk_dsi *dsi)
 		return;
 
 	regmap_update_bits(dsi->mmsys_sw_rst_b, dsi->sw_rst_b,
-			   MMSYS_SW_RST_DSI_B, ~MMSYS_SW_RST_DSI_B);
+			   MMSYS_SW_RST_DSI_B, (u32)~MMSYS_SW_RST_DSI_B);
 	usleep_range(1000, 1100);
 
 	regmap_update_bits(dsi->mmsys_sw_rst_b, dsi->sw_rst_b,
